aendret i DSL test, robot og wrapper

+++ Kode/SIR/ControlSystem/wrapper.cs
@@ -443,7 +443,7 @@ namespace ControlSystem
         public bool teachWrapped(SIRVector vecTheSirVector)
         {
             int iReturn;
-            for (int i = 1; i < vecTheSirVector.getSize(); i++)
+            for (short i = 1; i < vecTheSirVector.getSize(); i++)
             {
                 VecPoint pTmp = vecTheSirVector.getPoint(i-1);
                 int x, y, z, pitch, roll;
@@ -454,7 +454,7 @@ namespace ControlSystem
                 roll = pTmp.iRoll;
                 int[] iArray = new int[]{x, y, z, pitch, roll};
 
-                iReturn = _dll.Teach(vecTheSirVector.Name, (short) i, iArray, 5, vecTheSirVector.Type); // 8 because of 8 axes
+                iReturn = _dll.Teach(vecTheSirVector.Name, i, iArray, 5, vecTheSirVector.Type); // 8 because of 8 axes
                 if (iReturn == 0)
                     return (false);
             }
